+++ src/app/(actions)/issue.action.tsx
@@ -12,7 +12,7 @@ import {
 import { issueSearchListOutputValidator } from "~/app/(models)/dto/issue-search-output-validator";
 import { searchIssues } from "~/app/(models)/issues/search";
 
-import { getAuthedUser } from "./auth";
+import { getAuthedUser } from "./auth.action";
 import { cache } from "react";
 
 import type { IssueSearchFilters } from "~/lib/shared/utils.shared";
