+++ src/app/(components)/toast/toaster.server.tsx
@@ -1,7 +1,7 @@
 import * as React from "react";
 
 // utils
-import { getSession } from "~/app/(actions)/auth";
+import { getSession } from "~/app/(actions)/auth.action";
 import { ToasterClient } from "./toaster.client";
 import { headers } from "next/headers";
