+++ src/app/(routes)/(app)/settings/page.tsx
@@ -1,5 +1,5 @@
 import { redirect } from "next/navigation";
-import { redirectIfNotAuthed } from "~/app/(actions)/auth";
+import { redirectIfNotAuthed } from "~/app/(actions)/auth.action";
 
 export default async function Page() {
   await redirectIfNotAuthed("/settings/account");
